chore: resolve all PR suggestion #3242
